requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.30
python-dotenv~=1.0
pydantic~=2.9

